Fix Py3.11+ find_module not using given path
importlib.util.find_spec, as stated in the comment, doesn't take a path
but rather a package name. And this will be ignored unless name name
parameter starts with ".". Rather, So it will not search the given path,
rather it searches sys.meta_path.

Use importlib.machinery.PathFinder.find_spec instead, which searches in
the given path.